requirements.txt
Django>=5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==11.3.0
python-decouple==3.8
mysqlclient==2.2.7
django-extensions==3.2.3
whitenoise==6.6.0
django-filter==25.2